updates to products

+++ groundroots-admin/src/app/shared/services/product/product.service.ts
@@ -21,11 +21,8 @@ export class ProductService {
     return this.firestore.collection<Coffee>('products').doc(id).get();
   }
 
-  async create(product: Coffee, image: File) {
-    const doc = await this.firestore.collection('products').add(product);
-    const reference = this.storage.ref(`/products/${doc.id}`);
-    console.log(product)
-    return reference.put(image);
+  create(product: Coffee) {
+    return this.firestore.collection('products').add(product);
   }
 
   fetchProductImage(product: Coffee) {
@@ -38,7 +35,7 @@ export class ProductService {
 
   updateImage(product: Coffee, image: File) {
     const reference = this.storage.ref(`/products/${product.id}`);
-    reference.delete();
+    if (reference) reference.delete();
     return reference.put(image);
   }
 }
